PR fixes to CustomNavigationBar
