+++ WeatherApp/WeatherApp/Feature/Components/UINavigationController+BackSwipe.swift
@@ -1,0 +1,15 @@
+import SwiftUI
+
+extension UINavigationController: @retroactive UIGestureRecognizerDelegate {
+
+    override open func viewDidLoad() {
+        super.viewDidLoad()
+
+        interactivePopGestureRecognizer?.delegate = self
+    }
+
+    public func gestureRecognizerShouldBegin(_ gestureRecognizer: UIGestureRecognizer) -> Bool {
+        viewControllers.count > 1
+    }
+
+}
